Ulises

+++ TFGapp/TFGapp/Utility.cpp
@@ -3,7 +3,7 @@
 Horario * UTILITY existeHorario(vector<Horario> * horarios, string nombre, string dia)
 {
 	string nombreCompleto;
-	for (int i = 0; i < horarios->size(); i++)
+	for (int i = 0; i < (int)horarios->size(); i++)
 	{
 		nombreCompleto = (*horarios)[i].getProfesor().getNombre();
 		if (nombreCompleto == nombre && to_string((*horarios)[i].getDia()) == dia)
@@ -14,17 +14,16 @@ Horario * UTILITY existeHorario(vector<Horario> * horarios, string nombre, strin
 
 Profesor * UTILITY existeProfesor(vector<Profesor> *profesores, string ID)
 {
-	for (int i = 0; i < profesores->size(); i++)
+	for (int i = 0; i < (int)profesores->size(); i++)
 		if ((*profesores)[i].getNombre() == ID)
 			return &(*profesores)[i];
 	return NULL;
 }
 
-Grado* UTILITY existeGrado(vector<Grado> *grados, string ID)
+Grado * UTILITY existeGrado(vector<Grado> *grados, string ID)
 {
-	for (int i = 0; i < grados->size(); i++)
+	for (int i = 0; i < (int)grados->size(); i++)
 		if ((*grados)[i].getNombre() == ID)
 			return &(*grados)[i];
 	return NULL;
 }
-
